docs: Add section on out-of-tree modules to README
Added a new section in the README to explain "Out-of-tree
modules". This section details how modules can be loaded/unloaded,
assuming they follow the given specifications and are in the load
path. It also highlights projects that include their own tool
bindings through this library, such as "buffer-turtle" and
"arch-installer". This provides users with examples of integrations
with specific libraries. The update aims to help developers
understand how to extend the library with new modules.